delete and move stuff, convert discussion row
