--- client/scripts/views/pages/discussions/roles.ts
@@ -1,10 +1,0 @@
-import { RolePermission } from 'models';
-
-export const sortAdminsAndModsFirst = (a, b) => {
-  if (a.permission === b.permission) return a.Address.address.localeCompare(b.Address.address);
-  if (a.permission === RolePermission.admin) return -1;
-  if (b.permission === RolePermission.admin) return 1;
-  if (a.permission === RolePermission.moderator) return -1;
-  if (b.permission === RolePermission.moderator) return 1;
-  return a.Address.address.localeCompare(b.Address.address);
-};
