+++ client/scripts/views/pages/manage_community/index.ts
@@ -6,10 +6,19 @@ import $ from 'jquery';
 
 import app from 'state';
 import { RoleInfo, RolePermission, Webhook } from 'models';
-import { sortAdminsAndModsFirst } from 'views/pages/discussions/roles';
 import ChainMetadataManagementTable from './chain_metadata_management_table';
 import AdminPanelTabs from './admin_panel_tabs';
 import Sublayout from '../../sublayout';
+
+const sortAdminsAndModsFirst = (a, b) => {
+  if (a.permission === b.permission)
+    return a.Address.address.localeCompare(b.Address.address);
+  if (a.permission === RolePermission.admin) return -1;
+  if (b.permission === RolePermission.admin) return 1;
+  if (a.permission === RolePermission.moderator) return -1;
+  if (b.permission === RolePermission.moderator) return 1;
+  return a.Address.address.localeCompare(b.Address.address);
+};
 
 const ManageCommunityPage: m.Component<
   {},
